pyproject.toml
[project]
name = "bread-and-bureaucracy"
version = "0.1.0"
description = "A cave simulation game with a focus on bureaucracy."
dependencies = [
    "textual",
    "sqlmodel",
    "aiosqlite",
    "pyyaml",
    "pydantic",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]
